Move android native test execution to separate class
